[examples] consolidate summarization examples (#4837)
